refactor: configuration modules
