--- indexer/src/config/mod.rs
@@ -1,2 +1,0 @@
-pub mod config;
-pub mod user_config;
